Ensure each CFlow can be bound to VM scope on IOS (#15)
Current implementation of CFlow binds the coroutine scope for IOS to the
GlobalScope. This causes the observers on these flows to never close if
the ViewModel is disposed, since they are not part of the structured
concurrency. This could cause memory leaks as these resources will still
be open.

This is fixed by ensuring that the flows could also be bound by the VM
scope, and passing the ViewModelScope for the revolver view model usage
of CSharedFlow and CStateFlow. This will allow both global and VM scopes
to be used for CSharedFlow and CStateFlow if they require a Global scope
rather than being bound to the ViewModelScope.

See
[GlobalScope](https://kotlinlang.org/api/kotlinx.coroutines/kotlinx-coroutines-core/kotlinx.coroutines/-global-scope/)
for explanation why we have to bind the VM scope if we want the
observers to be part of the VM scope.