Update dependencies and gradle
Updating the dependencies on the project, also updating gradle to 8.11.1 to ensure that
the AGP of 8.7.0 can be ran without any issues. In the same time, updated the structure of the
Revolver build.gradle.properties to use the new androidLibrary block rather than
androidTarget and Android extension block. Also moved the kotlin plugin
from the gradle file to the gradle version catalog